requirements.txt
flask
flask-cors
python-docx
python-dotenv
huggingface-hub
gunicorn
pymongo
datetime
PyMuPDF